+++ lib/sshkit/formatters/dot.rb
@@ -4,13 +4,12 @@ module SSHKit
 
     class Dot < Abstract
 
-      def write(obj)
-        return unless obj.is_a? SSHKit::Command
-        if obj.finished?
-          original_output << colorize('.', obj.failure? ? :red : :green)
-        end
+      def log_command_exit(command)
+        original_output << colorize('.', command.failure? ? :red : :green)
       end
-      alias :<< :write
+
+      def write(obj)
+      end
 
     end