+++ lib/sshkit/formatters/pretty.rb
@@ -8,16 +8,35 @@ module SSHKit
       LEVEL_COLORS = [:black, :blue, :yellow, :red, :red].freeze
 
       def write(obj)
-        return if obj.respond_to?(:verbosity) && obj.verbosity < SSHKit.config.output_verbosity
-        case obj
-        when SSHKit::Command    then write_command(obj)
-        when SSHKit::LogMessage then write_message(obj.verbosity, obj.to_s)
+        if obj.kind_of?(SSHKit::LogMessage)
+          write_message(obj.verbosity, obj.to_s)
         else
-          raise "Output formatter only supports formatting SSHKit::Command and SSHKit::LogMessage, " \
-                "called with #{obj.class}: #{obj.inspect}"
+          raise "write only supports formatting SSHKit::LogMessage, called with #{obj.class}: #{obj.inspect}"
         end
       end
-      alias :<< :write
+
+      def log_command_start(command)
+        host_prefix = command.host.user ? "as #{colorize(command.host.user, :blue)}@" : 'on '
+        message = "Running #{colorize(command, :yellow, :bold)} #{host_prefix}#{colorize(command.host, :blue)}"
+        write_message(command.verbosity, message, command.uuid)
+        write_message(Logger::DEBUG, "Command: #{colorize(command.to_command, :blue)}", command.uuid)
+      end
+
+      def log_command_data(command, stream_type, stream_data)
+        color = case stream_type
+          when :stdout then :green
+          when :stderr then :red
+          else raise "Unrecognised stream_type #{stream_type}, expected :stdout or :stderr"
+        end
+        write_message(Logger::DEBUG, colorize("\t#{stream_data}".chomp, color), command.uuid)
+      end
+
+      def log_command_exit(command)
+        runtime = sprintf('%5.3f seconds', command.runtime)
+        successful_or_failed =  command.failure? ? colorize('failed', :red, :bold) : colorize('successful', :green, :bold)
+        message = "Finished in #{runtime} with exit status #{command.exit_status} (#{successful_or_failed})."
+        write_message(command.verbosity, message, command.uuid)
+      end
 
       protected
 
@@ -29,39 +48,8 @@ module SSHKit
 
       private
 
-      def write_command(command)
-        uuid = command.uuid
-
-        unless command.started?
-          host_prefix = command.host.user ? "as #{colorize(command.host.user, :blue)}@" : 'on '
-          message = "Running #{colorize(command, :yellow, :bold)} #{host_prefix}#{colorize(command.host, :blue)}"
-          write_message(command.verbosity, message, uuid)
-          write_debug("Command: #{colorize(command.to_command, :blue)}", uuid)
-        end
-
-        write_std_stream_debug(command.clear_stdout_lines, :green, uuid)
-        write_std_stream_debug(command.clear_stderr_lines, :red, uuid)
-
-        if command.finished?
-          runtime = sprintf('%5.3f seconds', command.runtime)
-          successful_or_failed =  command.failure? ? colorize('failed', :red, :bold) : colorize('successful', :green, :bold)
-          message = "Finished in #{runtime} with exit status #{command.exit_status} (#{successful_or_failed})."
-          write_message(command.verbosity, message, uuid)
-        end
-      end
-
-      def write_std_stream_debug(lines, color, uuid)
-        lines.each do |line|
-          write_debug(colorize("\t#{line}".chomp, color), uuid)
-        end
-      end
-
-      def write_debug(message, uuid)
-        write_message(Logger::DEBUG, message, uuid) if SSHKit.config.output_verbosity == Logger::DEBUG
-      end
-
       def write_message(verbosity, message, uuid=nil)
-        original_output << "#{format_message(verbosity, message, uuid)}\n"
+        original_output << "#{format_message(verbosity, message, uuid)}\n" if verbosity >= SSHKit.config.output_verbosity
       end
 
     end